requirements.txt
Django==1.6.5
Markdown==2.4.1
MySQL-python==1.2.5
argparse==1.2.1
django-filter==0.7
djangorestframework==2.3.14
pystache==0.5.4
wsgiref==0.1.2
